# Add type information to fix HssMatrix{T} when T is not a Float64