--- a/src/ulvfactor.jl
+++ b/src/ulvfactor.jl
@@ -11,7 +11,7 @@ function ulvfactsolve(hssA::HssMatrix{T}, b::Matrix{T}) where T
   if isleaf(hssA)
     return hssA.D\b
   else
-    z = zeros(eltype(hssA),size(hssA,2), size(b,2))
+    z = zeros(T, size(hssA,2), size(b,2))
     _, _, _, _, _, _, _, QV = _ulvfactsolve!(hssA, b, z, 0; rootnode=true)
     z = _ulvfactsolve_topdown!(QV, z)
     return z