Several functions were generating errors when called with HssMatrix{T}
 where T was not a Float64. These include: \ , / and compress_adatative.

This commit solves this problem and adds a test for it.