@bonevbs 2021-05-05
Hi Baptiste, thanks for contributing! I must admit `HssMatrices.jl` was written in quite a rush as I needed something to work with.

I will try to have a look soon so we can merge it :)