@BaptisteLamic 2021-05-06
You are welcome!  That’s a nice library. 
The test failed, but they were passing on my lab computer. It may due to the use of MKL instead of OpenBLAS, I will try to get them pass with the latter.