@BaptisteLamic 2022-12-22
Hi, this last version seems ready to be integrated for me. 
It would be nice that you could have a look.