@bonevbs 2022-12-22
Hi Baptiste, apologies for the slow replies. I have been quite busy with some other projects.

I just had a look, the code looks mostly good to me - please have a look at the comment I made.