@BaptisteLamic 2022-12-23
Hi, I just push a commit to add tests for random access. I will quickly have a look for performance regression.