@BaptisteLamic 2022-12-23
I just run the benchmarks on an Apple Silicon.  There are no notable differences. 

**New results**
Benchmarking full...
  13.976 ms (738 allocations: 70.91 MiB)
Benchmarking getindex...
  101.208 μs (1100 allocations: 711.03 KiB)
Benchmarking compression...
  124.240 ms (7775 allocations: 438.17 MiB)
Benchmarking randomized compression...
  34.689 ms (9514 allocations: 65.64 MiB)
Benchmarking re-compression...
  2.274 ms (9150 allocations: 4.82 MiB)
Benchmarking proper...
  3.309 ms (3946 allocations: 6.40 MiB)
Benchmarking addition...
  199.000 μs (545 allocations: 4.90 MiB)
Benchmarking matvec...
  340.625 μs (581 allocations: 1.47 MiB)
Benchmarking ulvfactsolve...
  8.547 ms (3259 allocations: 22.62 MiB)
Benchmarking hssldivide...
  32.619 ms (43121 allocations: 72.04 MiB)

**Old results** 
Benchmarking full...
  14.281 ms (738 allocations: 70.91 MiB)
Benchmarking getindex...
  106.583 μs (1100 allocations: 711.03 KiB)
Benchmarking compression...
  135.512 ms (7775 allocations: 438.17 MiB)
Benchmarking randomized compression...
  34.507 ms (9514 allocations: 65.64 MiB)
Benchmarking re-compression...
  2.213 ms (9150 allocations: 4.82 MiB)
Benchmarking proper...
  3.282 ms (3946 allocations: 6.40 MiB)
Benchmarking addition...
  198.875 μs (545 allocations: 4.90 MiB)
Benchmarking matvec...
  345.667 μs (581 allocations: 1.47 MiB)
Benchmarking ulvfactsolve...
  8.623 ms (3259 allocations: 22.62 MiB)
Benchmarking hssldivide...
  32.173 ms (43121 allocations: 72.04 MiB)