@bonevbs 2022-12-23
Thank you for testing the perf regression and all your work. I currently am quite busy and won't have time to check. Once you address my comment let us merge this :)